Fix stack frame issue.
We never cons a stack frame since we don't support calls (aside from the
initial call which doesn't push a stack frame); removing the cons
constructor solves the issue regarding memory separation.

This means we now _can't_ support calls even if we wanted to, but due to
the way we implement memory, we would need quite a lot of extra work to
support this.