Separate HTL instantiations from Verilog ones
In HTL, they reference their data arguments, destination, and finished
control signal. Meanwhile, in Verilog, they just contain an unstructured
list of parameters.

This is done because when modules are generated in the HTL stage we do
not have access to any of the instantiating module's control
signals (they are declared after the entirety of the module has been
translated). Also, I believe, these signals are not part of the HTL semantics.